test(node): split main tests into domain modules (#3605)
