+++ crates/kamn-node/src/main_tests.rs
@@ -215,6 +215,9 @@ fn spawn_kolme_live_mock_server(replies: Vec<MockHttpReply>) -> (String, Arc<Mut
 mod async_runtime_contract_tests;
 mod cli_contract_tests;
 mod core_behavior_tests;
+mod daemon_tests;
 mod observability_endpoint_tests;
+mod report_tests;
+mod runtime_tests;
 mod service_api_endpoint_tests;
 mod signer_tests;
