+++ crates/kamn-node/src/main_tests/daemon_tests.rs
@@ -1,0 +1,557 @@
+use super::*;
+
+#[cfg(unix)]
+const SIGTERM: i32 = 15;
+
+#[cfg(unix)]
+unsafe extern "C" {
+    fn raise(sig: i32) -> i32;
+}
+
+#[test]
+fn functional_runtime_daemon_emits_structured_transition_markers() {
+    let _lock = log_env_lock()
+        .lock()
+        .expect("log env lock should guard test mutation");
+    let _level_guard = EnvVarGuard::set("KAMN_NODE_LOG_LEVEL", Some("info"));
+    let _format_guard = EnvVarGuard::set("KAMN_NODE_LOG_FORMAT", Some("json"));
+    let parsed = parse_args(vec![
+        "kamn-node".to_owned(),
+        "--role".to_owned(),
+        "processor".to_owned(),
+        "--runtime-mode".to_owned(),
+        "daemon".to_owned(),
+        "--daemon-max-ticks".to_owned(),
+        "3".to_owned(),
+        "--daemon-tick-interval-ms".to_owned(),
+        "25".to_owned(),
+    ])
+    .expect("daemon args should parse");
+
+    let (report_result, captured_logs) = capture_test_logs(|| execute(parsed));
+    let report = report_result.expect("daemon execution should succeed");
+    assert_eq!(report.runtime_mode, "daemon");
+
+    let start_line = captured_logs
+        .iter()
+        .find(|line| line.contains("\"event\":\"node.runtime.daemon.execute.start\""))
+        .expect("daemon execution should emit structured start marker");
+    assert_eq!(
+        extract_json_string_field(start_line, "runtime_mode").as_deref(),
+        Some("daemon")
+    );
+    assert_eq!(
+        extract_json_string_field(start_line, "max_ticks").as_deref(),
+        Some("3")
+    );
+    assert_eq!(
+        extract_json_string_field(start_line, "tick_interval_ms").as_deref(),
+        Some("25")
+    );
+    let start_execution_id = extract_json_string_field(start_line, "execution_id")
+        .expect("daemon start marker should include execution_id");
+
+    let complete_line = captured_logs
+        .iter()
+        .find(|line| line.contains("\"event\":\"node.runtime.daemon.execute.complete\""))
+        .expect("daemon execution should emit structured completion marker");
+    assert_eq!(
+        extract_json_string_field(complete_line, "runtime_mode").as_deref(),
+        Some("daemon")
+    );
+    assert_eq!(
+        extract_json_string_field(complete_line, "executed_ticks").as_deref(),
+        Some("3")
+    );
+    assert_eq!(
+        extract_json_string_field(complete_line, "completion_reason").as_deref(),
+        Some("tick-budget-exhausted")
+    );
+    assert_eq!(
+        extract_json_string_field(complete_line, "shutdown_drain_status").as_deref(),
+        Some("not-signaled")
+    );
+    assert_eq!(
+        extract_json_string_field(complete_line, "shutdown_snapshot_flush_status").as_deref(),
+        Some("snapshot-not-requested")
+    );
+    assert_eq!(
+        extract_json_string_field(complete_line, "shutdown_signal_tick").as_deref(),
+        Some("none")
+    );
+    assert_eq!(
+        extract_json_string_field(complete_line, "shutdown_drain_ticks").as_deref(),
+        Some("0")
+    );
+    assert_eq!(
+        extract_json_string_field(complete_line, "shutdown_timeout_ticks").as_deref(),
+        Some("0")
+    );
+    assert_eq!(
+        extract_json_string_field(complete_line, "shutdown_ignored_signals").as_deref(),
+        Some("0")
+    );
+    let complete_execution_id = extract_json_string_field(complete_line, "execution_id")
+        .expect("daemon completion marker should include execution_id");
+    assert_eq!(start_execution_id, complete_execution_id);
+}
+
+#[test]
+fn functional_runtime_daemon_graceful_shutdown_emits_structured_drain_markers() {
+    let _lock = log_env_lock()
+        .lock()
+        .expect("log env lock should guard test mutation");
+    let _level_guard = EnvVarGuard::set("KAMN_NODE_LOG_LEVEL", Some("info"));
+    let _format_guard = EnvVarGuard::set("KAMN_NODE_LOG_FORMAT", Some("json"));
+    let parsed = parse_args(vec![
+        "kamn-node".to_owned(),
+        "--role".to_owned(),
+        "processor".to_owned(),
+        "--runtime-mode".to_owned(),
+        "daemon".to_owned(),
+        "--daemon-max-ticks".to_owned(),
+        "10".to_owned(),
+        "--daemon-tick-interval-ms".to_owned(),
+        "25".to_owned(),
+        "--daemon-shutdown-signal-tick".to_owned(),
+        "3".to_owned(),
+        "--daemon-shutdown-drain-ticks".to_owned(),
+        "2".to_owned(),
+        "--daemon-shutdown-timeout-ticks".to_owned(),
+        "4".to_owned(),
+    ])
+    .expect("daemon args should parse");
+
+    let (report_result, captured_logs) = capture_test_logs(|| execute(parsed));
+    let report = report_result.expect("daemon execution should succeed");
+    assert_eq!(report.runtime_mode, "daemon");
+
+    let complete_line = captured_logs
+        .iter()
+        .find(|line| line.contains("\"event\":\"node.runtime.daemon.execute.complete\""))
+        .expect("daemon execution should emit structured completion marker");
+    assert_eq!(
+        extract_json_string_field(complete_line, "completion_reason").as_deref(),
+        Some("graceful-shutdown:signal@3;drain_ticks=2;timeout_ticks=4;ignored_signals=0")
+    );
+    assert_eq!(
+        extract_json_string_field(complete_line, "shutdown_drain_status").as_deref(),
+        Some("completed")
+    );
+    assert_eq!(
+        extract_json_string_field(complete_line, "shutdown_signal_tick").as_deref(),
+        Some("3")
+    );
+    assert_eq!(
+        extract_json_string_field(complete_line, "shutdown_drain_ticks").as_deref(),
+        Some("2")
+    );
+    assert_eq!(
+        extract_json_string_field(complete_line, "shutdown_timeout_ticks").as_deref(),
+        Some("4")
+    );
+    assert_eq!(
+        extract_json_string_field(complete_line, "shutdown_ignored_signals").as_deref(),
+        Some("0")
+    );
+    assert_eq!(
+        extract_json_string_field(complete_line, "shutdown_snapshot_flush_status").as_deref(),
+        Some("snapshot-flushed")
+    );
+}
+
+#[test]
+pub(super) fn regression_runtime_daemon_shutdown_timeout_emits_structured_timeout_drain_markers() {
+    let _lock = log_env_lock()
+        .lock()
+        .expect("log env lock should guard test mutation");
+    let _level_guard = EnvVarGuard::set("KAMN_NODE_LOG_LEVEL", Some("info"));
+    let _format_guard = EnvVarGuard::set("KAMN_NODE_LOG_FORMAT", Some("json"));
+    let parsed = parse_args(vec![
+        "kamn-node".to_owned(),
+        "--role".to_owned(),
+        "processor".to_owned(),
+        "--runtime-mode".to_owned(),
+        "daemon".to_owned(),
+        "--daemon-max-ticks".to_owned(),
+        "10".to_owned(),
+        "--daemon-tick-interval-ms".to_owned(),
+        "25".to_owned(),
+        "--daemon-shutdown-signal-tick".to_owned(),
+        "7".to_owned(),
+        "--daemon-shutdown-drain-ticks".to_owned(),
+        "4".to_owned(),
+        "--daemon-shutdown-timeout-ticks".to_owned(),
+        "2".to_owned(),
+    ])
+    .expect("daemon timeout args should parse");
+
+    let (report_result, captured_logs) = capture_test_logs(|| execute(parsed));
+    let report = report_result.expect("daemon timeout execution should succeed");
+    assert_eq!(report.runtime_mode, "daemon");
+
+    let complete_line = captured_logs
+        .iter()
+        .find(|line| line.contains("\"event\":\"node.runtime.daemon.execute.complete\""))
+        .expect("daemon execution should emit structured completion marker");
+    assert_eq!(
+        extract_json_string_field(complete_line, "completion_reason").as_deref(),
+        Some("graceful-shutdown-timeout:signal@7;drain_ticks=4;timeout_ticks=2;ignored_signals=0")
+    );
+    assert_eq!(
+        extract_json_string_field(complete_line, "shutdown_drain_status").as_deref(),
+        Some("timeout")
+    );
+    assert_eq!(
+        extract_json_string_field(complete_line, "shutdown_signal_tick").as_deref(),
+        Some("7")
+    );
+    assert_eq!(
+        extract_json_string_field(complete_line, "shutdown_drain_ticks").as_deref(),
+        Some("4")
+    );
+    assert_eq!(
+        extract_json_string_field(complete_line, "shutdown_timeout_ticks").as_deref(),
+        Some("2")
+    );
+    assert_eq!(
+        extract_json_string_field(complete_line, "shutdown_ignored_signals").as_deref(),
+        Some("0")
+    );
+    assert_eq!(
+        extract_json_string_field(complete_line, "shutdown_snapshot_flush_status").as_deref(),
+        Some("snapshot-flush-timeout")
+    );
+}
+
+#[test]
+fn parses_runtime_mode_daemon_with_bounded_controls() {
+    let args = vec![
+        "kamn-node".to_owned(),
+        "--role".to_owned(),
+        "processor".to_owned(),
+        "--runtime-mode".to_owned(),
+        "daemon".to_owned(),
+        "--daemon-max-ticks".to_owned(),
+        "3".to_owned(),
+        "--daemon-tick-interval-ms".to_owned(),
+        "25".to_owned(),
+        "--daemon-peer-id".to_owned(),
+        "peer-alpha".to_owned(),
+        "--daemon-lifecycle-event".to_owned(),
+        "start-connect".to_owned(),
+        "--daemon-lifecycle-event".to_owned(),
+        "handshake-succeeded".to_owned(),
+    ];
+
+    let parsed = parse_args(args).expect("daemon args should parse");
+    assert_eq!(parsed.runtime_mode, RuntimeMode::daemon());
+    assert_eq!(parsed.daemon_max_ticks, Some(3));
+    assert_eq!(parsed.daemon_tick_interval_ms, Some(25));
+    assert!(!parsed.daemon_shutdown_os_signals);
+    assert_eq!(parsed.daemon_peer_id, Some("peer-alpha".to_owned()));
+    assert_eq!(parsed.daemon_lifecycle_events.len(), 2);
+}
+
+#[test]
+fn parses_runtime_mode_daemon_with_shutdown_controls() {
+    let args = vec![
+        "kamn-node".to_owned(),
+        "--role".to_owned(),
+        "processor".to_owned(),
+        "--runtime-mode".to_owned(),
+        "daemon".to_owned(),
+        "--daemon-max-ticks".to_owned(),
+        "8".to_owned(),
+        "--daemon-tick-interval-ms".to_owned(),
+        "25".to_owned(),
+        "--daemon-shutdown-signal-tick".to_owned(),
+        "3".to_owned(),
+        "--daemon-shutdown-drain-ticks".to_owned(),
+        "2".to_owned(),
+        "--daemon-shutdown-timeout-ticks".to_owned(),
+        "4".to_owned(),
+    ];
+
+    let parsed = parse_args(args).expect("daemon args with shutdown controls should parse");
+    assert_eq!(parsed.daemon_shutdown_signal_ticks, vec![3]);
+    assert!(!parsed.daemon_shutdown_os_signals);
+    assert_eq!(parsed.daemon_shutdown_drain_ticks, Some(2));
+    assert_eq!(parsed.daemon_shutdown_timeout_ticks, Some(4));
+}
+
+#[test]
+fn parses_runtime_mode_daemon_with_os_signal_shutdown_controls() {
+    let args = vec![
+        "kamn-node".to_owned(),
+        "--role".to_owned(),
+        "processor".to_owned(),
+        "--runtime-mode".to_owned(),
+        "daemon".to_owned(),
+        "--daemon-max-ticks".to_owned(),
+        "12".to_owned(),
+        "--daemon-tick-interval-ms".to_owned(),
+        "5".to_owned(),
+        "--daemon-shutdown-os-signals".to_owned(),
+        "--daemon-shutdown-drain-ticks".to_owned(),
+        "2".to_owned(),
+        "--daemon-shutdown-timeout-ticks".to_owned(),
+        "4".to_owned(),
+    ];
+
+    let parsed = parse_args(args).expect("daemon args with os signal controls should parse");
+    assert_eq!(parsed.daemon_shutdown_signal_ticks, Vec::<u64>::new());
+    assert!(parsed.daemon_shutdown_os_signals);
+    assert_eq!(parsed.daemon_shutdown_drain_ticks, Some(2));
+    assert_eq!(parsed.daemon_shutdown_timeout_ticks, Some(4));
+}
+
+#[test]
+fn parses_runtime_mode_daemon_with_observability_endpoint_controls() {
+    let args = vec![
+        "kamn-node".to_owned(),
+        "--role".to_owned(),
+        "processor".to_owned(),
+        "--runtime-mode".to_owned(),
+        "daemon".to_owned(),
+        "--daemon-max-ticks".to_owned(),
+        "12".to_owned(),
+        "--daemon-tick-interval-ms".to_owned(),
+        "5".to_owned(),
+        "--observability-endpoint-bind".to_owned(),
+        "127.0.0.1:9108".to_owned(),
+        "--observability-endpoint-metrics-path".to_owned(),
+        "/runtime/metrics".to_owned(),
+        "--observability-endpoint-health-path".to_owned(),
+        "/runtime/health".to_owned(),
+        "--observability-endpoint-max-requests".to_owned(),
+        "3".to_owned(),
+        "--observability-endpoint-idle-timeout-ms".to_owned(),
+        "1200".to_owned(),
+    ];
+
+    let parsed = parse_args(args).expect("daemon args with observability endpoint should parse");
+    assert_eq!(
+        parsed.observability_endpoint_bind_addr,
+        Some("127.0.0.1:9108".to_owned())
+    );
+    assert_eq!(
+        parsed.observability_endpoint_metrics_path,
+        "/runtime/metrics"
+    );
+    assert_eq!(parsed.observability_endpoint_health_path, "/runtime/health");
+    assert_eq!(parsed.observability_endpoint_max_requests, 3);
+    assert_eq!(parsed.observability_endpoint_idle_timeout_ms, 1200);
+}
+
+#[test]
+fn env_only_daemon_controls_parse_without_config_file() {
+    let _env_lock = signer_env_lock()
+        .lock()
+        .expect("env lock should guard process-level overrides");
+    let _max_ticks_guard = EnvVarGuard::set("KAMN_NODE_DAEMON_MAX_TICKS", Some("12"));
+    let _tick_interval_guard = EnvVarGuard::set("KAMN_NODE_DAEMON_TICK_INTERVAL_MS", Some("25"));
+
+    let args = vec![
+        "kamn-node".to_owned(),
+        "--role".to_owned(),
+        "processor".to_owned(),
+        "--runtime-mode".to_owned(),
+        "daemon".to_owned(),
+    ];
+
+    let parsed = parse_args(args).expect("env-only daemon controls should parse");
+    assert_eq!(parsed.daemon_max_ticks, Some(12));
+    assert_eq!(parsed.daemon_tick_interval_ms, Some(25));
+}
+
+#[test]
+fn regression_3202_invalid_daemon_env_override_fails_closed_without_config_file() {
+    let _env_lock = signer_env_lock()
+        .lock()
+        .expect("env lock should guard process-level overrides");
+    let _max_ticks_guard = EnvVarGuard::set("KAMN_NODE_DAEMON_MAX_TICKS", Some("invalid"));
+    let _tick_interval_guard = EnvVarGuard::set("KAMN_NODE_DAEMON_TICK_INTERVAL_MS", Some("25"));
+
+    let args = vec![
+        "kamn-node".to_owned(),
+        "--role".to_owned(),
+        "processor".to_owned(),
+        "--runtime-mode".to_owned(),
+        "daemon".to_owned(),
+    ];
+
+    let parse_result = parse_args(args);
+    assert!(
+        matches!(
+            parse_result,
+            Err(ConfigError::InvalidDaemonControlArgument(value)) if value == "invalid"
+        ),
+        "invalid daemon env override must fail closed with typed daemon control error"
+    );
+}
+
+#[test]
+fn integration_runtime_daemon_renders_bounded_completion_output() {
+    let args = vec![
+        "kamn-node".to_owned(),
+        "--role".to_owned(),
+        "processor".to_owned(),
+        "--runtime-mode".to_owned(),
+        "daemon".to_owned(),
+        "--daemon-max-ticks".to_owned(),
+        "3".to_owned(),
+        "--daemon-tick-interval-ms".to_owned(),
+        "25".to_owned(),
+        "--daemon-peer-id".to_owned(),
+        "peer-alpha".to_owned(),
+        "--daemon-lifecycle-event".to_owned(),
+        "start-connect".to_owned(),
+        "--daemon-lifecycle-event".to_owned(),
+        "handshake-succeeded".to_owned(),
+        "--daemon-lifecycle-event".to_owned(),
+        "heartbeat-missed".to_owned(),
+        "--daemon-lifecycle-event".to_owned(),
+        "heartbeat-restored".to_owned(),
+        "--output".to_owned(),
+        "json".to_owned(),
+    ];
+
+    let parsed = parse_args(args).expect("daemon args should parse");
+    let report = execute(parsed).expect("daemon execution should succeed");
+    let rendered = render_bootstrap_report(&report, OutputMode::json());
+    assert!(rendered.contains("\"runtime_mode\":\"daemon\""));
+    assert!(rendered.contains("\"daemon_max_ticks\":3"));
+    assert!(rendered.contains("\"daemon_tick_interval_ms\":25"));
+    assert!(rendered.contains("\"daemon_executed_ticks\":3"));
+    assert!(rendered.contains("\"daemon_completion_reason\":\"tick-budget-exhausted\""));
+    assert!(rendered.contains("\"daemon_observability_latency_p50_ms\":25"));
+    assert!(rendered.contains("\"daemon_observability_latency_p99_ms\":50"));
+    assert!(rendered.contains("\"daemon_observability_throughput_tps\":2000"));
+    assert!(rendered.contains("\"daemon_observability_error_rate_bps\":50"));
+    assert!(rendered.contains("\"daemon_observability_availability_bps\":9990"));
+    assert!(rendered.contains("\"daemon_observability_health\":\"healthy\""));
+    assert!(rendered.contains("\"daemon_observability_alert_count\":0"));
+    assert!(rendered.contains("\"daemon_observability_reason_code\":\"none\""));
+    assert!(rendered.contains("\"daemon_observability_transport_checkpoint_failures\":0"));
+    assert!(rendered.contains("\"daemon_observability_signer_checkpoint_failures\":0"));
+    assert!(rendered.contains("\"daemon_observability_commit_checkpoint_failures\":0"));
+    assert!(rendered.contains("\"daemon_peer_id\":\"peer-alpha\""));
+    assert!(rendered.contains("\"daemon_peer_lifecycle_final_state\":\"active\""));
+    assert!(
+        rendered.contains(
+            "\"daemon_peer_lifecycle_applied_events\":[\"start-connect\",\"handshake-succeeded\",\"heartbeat-missed\",\"heartbeat-restored\"]"
+        )
+    );
+}
+
+#[test]
+fn functional_runtime_daemon_applies_graceful_shutdown_signal() {
+    let args = vec![
+        "kamn-node".to_owned(),
+        "--role".to_owned(),
+        "processor".to_owned(),
+        "--runtime-mode".to_owned(),
+        "daemon".to_owned(),
+        "--daemon-max-ticks".to_owned(),
+        "10".to_owned(),
+        "--daemon-tick-interval-ms".to_owned(),
+        "25".to_owned(),
+        "--daemon-shutdown-signal-tick".to_owned(),
+        "3".to_owned(),
+        "--daemon-shutdown-drain-ticks".to_owned(),
+        "2".to_owned(),
+        "--daemon-shutdown-timeout-ticks".to_owned(),
+        "4".to_owned(),
+        "--output".to_owned(),
+        "json".to_owned(),
+    ];
+
+    let parsed = parse_args(args).expect("daemon shutdown args should parse");
+    let report = execute(parsed).expect("daemon graceful shutdown execution should succeed");
+    let rendered = render_bootstrap_report(&report, OutputMode::json());
+    assert!(rendered.contains("\"daemon_executed_ticks\":5"));
+    assert!(rendered.contains(
+        "\"daemon_completion_reason\":\"graceful-shutdown:signal@3;drain_ticks=2;timeout_ticks=4;ignored_signals=0\""
+    ));
+    assert!(rendered.contains("\"daemon_observability_health\":\"healthy\""));
+    assert!(rendered.contains("\"daemon_observability_alert_count\":0"));
+}
+
+#[test]
+fn integration_runtime_daemon_shutdown_timeout_is_fail_closed() {
+    let args = vec![
+        "kamn-node".to_owned(),
+        "--role".to_owned(),
+        "processor".to_owned(),
+        "--runtime-mode".to_owned(),
+        "daemon".to_owned(),
+        "--daemon-max-ticks".to_owned(),
+        "10".to_owned(),
+        "--daemon-tick-interval-ms".to_owned(),
+        "25".to_owned(),
+        "--daemon-shutdown-signal-tick".to_owned(),
+        "7".to_owned(),
+        "--daemon-shutdown-drain-ticks".to_owned(),
+        "4".to_owned(),
+        "--daemon-shutdown-timeout-ticks".to_owned(),
+        "2".to_owned(),
+        "--output".to_owned(),
+        "json".to_owned(),
+    ];
+
+    let parsed = parse_args(args).expect("daemon timeout args should parse");
+    let report = execute(parsed).expect("daemon timeout execution should succeed");
+    let rendered = render_bootstrap_report(&report, OutputMode::json());
+    assert!(rendered.contains("\"daemon_executed_ticks\":9"));
+    assert!(rendered.contains(
+        "\"daemon_completion_reason\":\"graceful-shutdown-timeout:signal@7;drain_ticks=4;timeout_ticks=2;ignored_signals=0\""
+    ));
+    assert!(rendered.contains("\"daemon_observability_latency_p50_ms\":145"));
+    assert!(rendered.contains("\"daemon_observability_latency_p99_ms\":425"));
+    assert!(rendered.contains("\"daemon_observability_throughput_tps\":900"));
+    assert!(rendered.contains("\"daemon_observability_error_rate_bps\":250"));
+    assert!(rendered.contains("\"daemon_observability_availability_bps\":9800"));
+    assert!(rendered.contains("\"daemon_observability_health\":\"critical\""));
+    assert!(rendered.contains("\"daemon_observability_alert_count\":4"));
+    assert!(rendered.contains("\"daemon_observability_reason_code\":\"daemon_shutdown_timeout\""));
+    assert!(rendered.contains("\"daemon_observability_transport_checkpoint_failures\":0"));
+    assert!(rendered.contains("\"daemon_observability_signer_checkpoint_failures\":0"));
+    assert!(rendered.contains("\"daemon_observability_commit_checkpoint_failures\":1"));
+}
+
+#[cfg(unix)]
+#[test]
+pub(super) fn integration_runtime_daemon_applies_graceful_shutdown_on_os_signal() {
+    let args = vec![
+        "kamn-node".to_owned(),
+        "--role".to_owned(),
+        "processor".to_owned(),
+        "--runtime-mode".to_owned(),
+        "daemon".to_owned(),
+        "--daemon-max-ticks".to_owned(),
+        "100".to_owned(),
+        "--daemon-tick-interval-ms".to_owned(),
+        "1".to_owned(),
+        "--daemon-shutdown-os-signals".to_owned(),
+        "--daemon-shutdown-drain-ticks".to_owned(),
+        "2".to_owned(),
+        "--daemon-shutdown-timeout-ticks".to_owned(),
+        "5".to_owned(),
+        "--output".to_owned(),
+        "json".to_owned(),
+    ];
+
+    let parsed = parse_args(args).expect("daemon os-signal args should parse");
+    let trigger = std::thread::spawn(|| {
+        std::thread::sleep(std::time::Duration::from_millis(10));
+        let result = unsafe { raise(SIGTERM) };
+        assert_eq!(result, 0, "test SIGTERM raise should succeed");
+    });
+    let report = execute(parsed).expect("daemon os-signal execution should succeed");
+    trigger
+        .join()
+        .expect("os-signal trigger thread should complete");
+    let rendered = render_bootstrap_report(&report, OutputMode::json());
+    assert!(rendered.contains("\"daemon_completion_reason\":\"graceful-shutdown:signal@"));
+}
